Let the yumconf.persistdir be as the default
This path gets appended to the installroot every time,
even if it's an absolute path, so it ends up being inside
the installroot everytime. We don't want it to be in some
path depending on a temporary workdir. This will put all
the yum files in a standard /var/lib/yum directory inside
the installroot.