Add support for --installpkgs
This allows the user to add specific packages, or package globs, to the
installer's root via lorax. For example, to build a server product with
the correct product.img you would pass --installpkgs
fedora-productimg-server

This removes the need for the kickstart to use --exclude on the repo
lines and makes it more explicit as to what is being built. This command
mirrors the same command in lorax.